svsCut/requirements.txt
imageio==2.33.1
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
networkx==3.2.1
numpy==1.26.4
openslide-python==1.3.1
packaging==24.1
pillow==10.3.0
pydicom==2.4.4
scikit-image==0.24.0
scipy==1.13.1
tifffile==2024.6.18
